requirements.txt
flask==3.0.3
gunicorn==23.0.0
tflite-runtime==2.11.0
Pillow==10.4.0
numpy==1.26.4
